requirements.txt
anthropic
requests
python-dotenv
flask
flask-cors
